requirements.txt
duckduckgo-search 
langchain-community
crawl4ai
langchain
openai
langchain_groq
duckduckgo-search
python-dotenv
pydantic
beautifulsoup4
readability-lxml
pytest
lxml[html_clean]
"fastapi[standard]" 
streamlit

# NOTE: Before running the project, update the .env file with your API key.
# Example:
# Groq_API_KEY=your_actual_api_key_here